requirements.txt
prefect
python
flask
flask_login
flask_sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
langchain
langchain-google-genai
langchain-groq
torch 
transformers 
accelerate 
sentencepiece
langchain_text_splitters
mistralai
langchain-mistralai
langchain-pinecone
tiktoken
sec-api
pymupdf
llama_parse
llama_cloud
toons